No longer using ntp server. Just using time MQTT topic
The time mqtt topic will be updated by the HA server every day. During a power outage, a UPS could keep the HA server alive so it can keep the time. The controllers will power off before the generator kicks in and lose their time. They will be able to resync with the HA server.